+++ utils/read-file.js
@@ -4,7 +4,6 @@ module.exports = (pathToFile) => {
   return fsPromises
     .readFile(pathToFile, { encoding: 'utf8' })
     .then((data) => {
-      console.log('data', data);
       return JSON.parse(data);
     })
     .catch((err) => err);
